requirements.txt
accelerate==1.12.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
exceptiongroup==1.3.1
filelock==3.20.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==12.0.0
psutil==7.2.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
shellingham==1.5.4
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1+cpu
tqdm==4.67.1
transformers==5.0.0rc1
typer-slim==0.20.1
typing_extensions==4.15.0
urllib3==2.6.2
